removed circular dependency
