+++ packages/web-sdk/src/instrumentations/userActions/userActionController.ts
@@ -5,9 +5,8 @@ import type { Subscription, UserActionInterface } from '@grafana/faro-core';
 import { monitorDomMutations } from './domMutationMonitor';
 import { monitorHttpRequests } from './httpRequestMonitor';
 import { monitorPerformanceEntries } from './performanceEntriesMonitor';
-import { isRequestEndMessage, isRequestStartMessage } from './processUserActionEventHandler';
 import type { HttpRequestMessagePayload } from './types';
-import { startTimeout } from './util';
+import { isRequestEndMessage, isRequestStartMessage, startTimeout } from './util';
 
 const defaultFollowUpActionTimeRange = 100;
 const defaultHaltTimeout = 10 * 1000;
